Improve subtitle rendering
* Go back to rendering subtitles with the UI text rendering widgets.
  Looks much better as it does not need to be scaled. Also modifications
  to font scale and position will take effect immediately.

* Improve text rendering so it can changes font face, color and size during
  a rendered text.

* Load fonts attached in MKV files into font rendering subsystem.

* Improve audio and subtitle settings in video playback pop-up menu.
  Reuse same setting model and widgets as the global settings menu.

IssueID #569
IssueID #540

+++ src/ui/glw/glw_array.c
@@ -487,7 +487,7 @@ glw_array_get_num_children_x(glw_t *w)
  *
  */
 static void
-set_margin(glw_t *w, const float *v)
+set_margin(glw_t *w, const int16_t *v)
 {
   glw_array_t *a = (glw_array_t *)w;
   a->margin_left   = v[0];
@@ -501,7 +501,7 @@ set_margin(glw_t *w, const float *v)
  *
  */
 static void
-set_border(glw_t *w, const float *v)
+set_border(glw_t *w, const int16_t *v)
 {
   glw_array_t *a = (glw_array_t *)w;
   a->border_left   = v[0];
